updated pot file + en.po & added german translation
This is my first time working of translation files. I hope I didn't mess anything up ;)
Just to be sure, I backed up the old versions by adding ".old".

I added the german translation, however currently it only works with my adjustments in the settings menu (making it dynamic) and not with the static string for each workday.

the czech translation still needs to be adjusted. However, the changes should be quite minor.