Do not take into account days that open and close at the same time
Hi,

Waiting for #87 to be addressed, I would suggest this small addition to disable days with same open and close time. Anyway it could be seen as something logical because if you open and close at the same time then you don't stay opened enough to allow any pick-up ;)

And thanks very much for this plugin !

Séb.