Markup inconsistency and duplicate sentence
Hi,

Here I replaced a `<h2>` with a `<h3>` to stay consistent with WooCommerce markup on checkout page.

I also replaced the second "Pickup Time" with "Select time", a string used elsewhere in the plugin, because as it is now you have twice the same sentence appearing one after the other : "Pickup Time" as a title then as a field label. And there is no other way to avoid this...

Thanks again !

Séb.